+++ tom_demark_indicator/__init__.py
@@ -4,6 +4,7 @@ from .indicators import add_indicators
 from .td_sequential import add_td_sequential
 from .plotting_mpf import plot_with_mplfinance
 from .exporter import save_data_json, default_image_path
+from .signals import run_daily_signals, get_stock_list
 
 __all__ = [
     "PlotConfig",
@@ -13,4 +14,6 @@ __all__ = [
     "plot_with_mplfinance",
     "save_data_json",
     "default_image_path",
+    "run_daily_signals",
+    "get_stock_list",
 ]